+++ src/Pages/Cart.tsx
@@ -1,0 +1,10 @@
+import {View, Text} from 'react-native';
+import React from 'react';
+
+export default function Cart() {
+  return (
+    <View>
+      <Text>Cart</Text>
+    </View>
+  );
+}
