+++ src/Pages/Product.tsx
@@ -1,0 +1,10 @@
+import {View, Text} from 'react-native';
+import React from 'react';
+
+export default function Product() {
+  return (
+    <View>
+      <Text>Product</Text>
+    </View>
+  );
+}
